+++ src/main/java/com/warmup/step1/StringHelper.java
@@ -1,14 +1,15 @@
-package requirement;
+package com.warmup.step1;
 
-public class StringHelper {
+class StringHelper {
 
     public static String[] splitter(String input) {
         if (input.startsWith("(")) {
             input = input.substring(1);
         }
-        
+
         if (input.endsWith(")")) {
             input = input.substring(0, input.length() - 1);
+
         }
 
         return input.split(",");
