+++ src/test/java/com/warmup/step1/SetTest.java
@@ -1,4 +1,4 @@
-package requirement;
+package com.warmup.step1;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
