+++ src/test/java/com/warmup/step1/StringTest.java
@@ -1,4 +1,4 @@
-package requirement;
+package com.warmup.step1;
 
 import org.junit.jupiter.api.Test;
